+++ cli/cli/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - `beam deploy` commands use solution level building instead of per-project [#3952](https://github.com/beamable/BeamableProduct/issues/3952)
-- `beam project open` command adds small delay before opening `.sln` to appease Visual Studio
+- `beam project open` command can create a `.slnf` file to show a subset of projects based on the Unity project perspective. 
 
 ### Added
 - `beam deploy release` shows CID/PID information as part of release confirmation [#3954](https://github.com/beamable/BeamableProduct/issues/3954)